# packages/sdk/openapi/agentflow.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentFlow API",
    "version": "0.1.0",
    "description": "REST API for AgentFlow — an open-source workflow automation platform.\n\n**Authentication** — send `Authorization: Bearer <token>` with a JWT access token (POST /api/auth/login) or a personal API key (`af_…`, POST /api/api-keys).\n\n**Rate limiting** — 200 requests/minute per user or IP by default; stricter limits apply to auth, AI and checkout endpoints.\n\n**Pagination** — list endpoints accept `page` and `limit` (max 100) and return the applied values in the `X-Page` and `X-Limit` response headers.",
    "license": {
      "name": "MIT"
    }
  },
  "paths": {
...
    "/api/executions/trigger": {
      "post": {
        "tags": [
          "Executions"
        ],
        "summary": "Trigger a workflow execution via API",
        "description": "Quota-checked. Ideal for scheduled/CI invocations using an API key.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Execution request",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "workflowId": {
                    "type": "string",
                    "minLength": 1,
                    "example": "clx4wf777"
                  },
                  "input": {
                    "type": "object",
                    "additionalProperties": {},
                    "description": "Initial execution input"
                  },
                  "trigger": {
                    "type": "string",
                    "enum": [
                      "manual",
                      "webhook",
                      "cron",
                      "api"
                    ],
                    "default": "api"
                  }
                },
                "required": [
                  "workflowId"
                ]
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Execution created and queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Execution"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Monthly execution quota exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Workflow not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/executions": {
      "get": {
        "tags": [
          "Executions"
        ],
        "summary": "List executions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "example": 1
            },
            "required": true,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "example": 25
            },
            "required": true,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "workflowId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "RUNNING",
                "SUCCESS",
                "FAILED",
                "CANCELLED",
                "WAITING_APPROVAL"
              ]
            },
            "required": false,
            "name": "status",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Executions, newest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/Execution"
                      },
                      {
                        "type": "object",
                        "properties": {
                          "workflow": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "name"
                            ]
                          }
                        },
                        "required": [
                          "workflow"
                        ]
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/executions/{id}": {
      "get": {
        "tags": [
          "Executions"
        ],
        "summary": "Get execution details (nodes, approvals, workflow)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Entity ID (cuid)"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Execution detail",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Execution"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "nodes": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/NodeExecution"
                          }
                        },
                        "approvals": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Approval"
                          }
                        },
                        "workflow": {
                          "type": [
                            "object",
                            "null"
                          ],
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "id",
                            "name"
                          ]
                        }
                      },
                      "required": [
                        "nodes",
                        "approvals",
                        "workflow"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Execution not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/executions/{id}/cancel": {
      "post": {
        "tags": [
          "Executions"
        ],
        "summary": "Cancel a pending or running execution",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Entity ID (cuid)"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found or already finished (NOT_CANCELLABLE)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/executions/{id}/nodes": {
      "get": {
        "tags": [
          "Executions"
        ],
        "summary": "List node-level logs of an execution",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Entity ID (cuid)"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Node executions in run order",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NodeExecution"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Execution not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "code": {
            "type": "string",
            "description": "Machine-readable error code",
            "example": "NOT_FOUND"
          },
          "details": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": "Field-level validation errors (VALIDATION_ERROR only)"
          }
        },
        "required": [
          "error",
          "code"
        ],
        "description": "Error envelope used by every non-2xx response"
      },
...
      "OkResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "ok"
        ]
      },
...
      "Execution": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "clx4ex321"
          },
          "workflowId": {
            "type": "string"
          },
          "orgId": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "RUNNING",
              "SUCCESS",
              "FAILED",
              "CANCELLED",
              "WAITING_APPROVAL"
            ],
            "example": "SUCCESS"
          },
          "trigger": {
            "type": "string",
            "enum": [
              "manual",
              "webhook",
              "cron",
              "api"
            ],
            "example": "manual"
          },
          "input": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {}
          },
          "output": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {}
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "finishedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "workflowId",
          "orgId",
          "status",
          "trigger",
          "input",
          "output",
          "error",
          "startedAt",
          "finishedAt"
        ]
      },
      "NodeExecution": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "executionId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "RUNNING",
              "SUCCESS",
              "FAILED",
              "SKIPPED"
            ],
            "example": "SUCCESS"
          },
          "input": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {}
          },
          "output": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {}
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "finishedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "executionId",
          "status",
          "input",
          "output",
          "error",
          "startedAt",
          "finishedAt"
        ]
      },
      "Approval": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "executionId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED"
            ],
            "example": "PENDING"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "decidedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "execution": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "workflow": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  }
                },
                "required": [
                  "id",
                  "name"
                ]
              }
            },
            "required": [
              "id",
              "workflow"
            ]
          }
        },
        "required": [
          "id",
          "executionId",
          "status",
          "createdAt",
          "decidedAt"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "AgentFlow access token (from POST /api/auth/login) or a personal API key (`af_…`, from POST /api/api-keys). Send as `Authorization: Bearer <token>`."
      },
      "webhookSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Webhook-Signature",
        "description": "HMAC-SHA256 hex digest of the raw request body, computed with the webhook secret (optional `sha256=` prefix). Required on every /api/webhooks/trigger/{path} call."
      },
      "stripeSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "stripe-signature",
        "description": "Stripe webhook signature header (t=…,v1=…)."
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development"
    },
    {
      "url": "https://api.agentflow.dev",
      "description": "Production (placeholder)"
    }
  ],
...
}